requirements.txt
# JARVIS Trading Agent System - Phase 1
# Core
openai>=1.0.0          # OpenAI-compatible client for NVIDIA NIM
ccxt>=4.0.0            # Exchange API (100+ exchanges)
python-dotenv>=1.0.0   # Environment variables

# Data & Analysis
pandas>=2.0.0          # Data manipulation
ta>=0.11.0             # Technical analysis indicators
numpy>=1.24.0          # Numerical computing

# Async & Scheduling
asyncio-atexit>=1.0.0  # Async cleanup
aiohttp>=3.9.0         # Async HTTP requests

# News & Feeds
feedparser>=6.0.0      # RSS feed parsing
beautifulsoup4>=4.12.0 # HTML parsing

# Interface
rich>=13.0.0           # Beautiful terminal output
